Completely accurate short range IR ADC and reads multiple IRs

+++ Milestones/Target_Grabber/firmware/src/adc_thread_public.h
@@ -23,9 +23,9 @@ extern "C" {
   
 // This struct contains the three IR sensor values
 typedef struct {
-    float IR_0;
-    float IR_1;
-    float IR_2;
+    uint32_t IR_0;
+    uint32_t IR_1;
+    uint32_t IR_2;
 } TargetAlignment_Type;
